+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - Assorted documentation updates
 - Add `bencapex` as illustrative data frame.
+- Add `bcai()` for calculating Benati and Capurri's (2026) alignment index. You can think of this as a kind of chance-corrected *S* score.
 
 # fpsim 0.1.0